+++ common/src/main/java/me/pandamods/pandalib/client/animation_controller/PlayType.java
@@ -1,0 +1,6 @@
+package me.pandamods.pandalib.client.animation_controller;
+
+public enum PlayType {
+	LOOP,
+	PLAY
+}